Move subscribe API tests to v2
