+++ tests/jest.config.integration.apiv2.sdkv5.js
@@ -18,8 +18,8 @@ module.exports = {
 	testMatch: [
 		'<rootDir>/integration/api_v2/http/*.test.js',
 		'<rootDir>/integration/api_v2/rpc/*.test.js',
+		'<rootDir>/integration/api_v2/subscribeApi/*.test.js',
 		'<rootDir>/integration/gateway/*.test.js',
-		'<rootDir>/integration/subscribeApi/*.test.js',
 	],
 	testEnvironment: 'node',
 	testTimeout: 50000,
